fix: create serivce with nonexistent variables
